Bug Fix #5: Create Template: Tri-state checkbox should appear in file tree for parents that have some children checked and some children unchecked.
Solution: Added a check parent that induces that checks its children to
see if which state it should be put in. Added in treemodelrestore and
treemodeluse as well.

Now allows filter model to change checked values as well in the
dictionary
Changed Package extension names to .ezpack
Omitted print statements in checkboxdelegate.py