Rewrote the backend that represents the "template" file using python dataclasses
Used the @DataClass to write a pure python backend model called the EzMetadataModel
that contains a list of EzMetadataEntry which is also a pure python data class. We
then used the dataclasses-json to serialize to a from JSON. These changes allow
the developer/user to push data to a HyperThought instance using both the GUI and
the programmatically in python. During this rewrite many of the classes that make
up the application needed modifications. This commit represents a feature parity
with the original work that ICS delivered to BlueQuartz/NRL.

The data model classes are in the 'ezmodel' python module.
Generated UI files start with the ui_* prefix
CSS Based Style Sheet assets were added but styling is disabled for now.
Packages are implemented.
Authenticating to a remote HyperThought instance to set the upload directory has
been updated with a few more features.
Some classes have been renamed along the way.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>