Added a generic XML parser
This is a first pass, and it has a few limitations principly around
dropping any duplicate tags. It will place the first tag in the model,
and drop any duplicates. Further work would be needed to add uniqueness
to tag names to map to the simple model expected by MetaForge.

Signed-off-by: Marcus D. Hanwell <marcus@quantumista.com>