Use a node with .attrs appended for attributes
Issue #61 outlined issues with the attributes as implemented. This
switches the attributes to always append '.attrs' to the path to the
attribute entries. This avoids the situation where there is data the
user wants to remove from their template but they want to keep some
attributes. This resolves the issue outlines as the problematic entry is
never emitted from the parser now.