Attributes (#74)
* Use a node with .attrs appended for attributes

Issue #61 outlined issues with the attributes as implemented. This
switches the attributes to always append '.attrs' to the path to the
attribute entries. This avoids the situation where there is data the
user wants to remove from their template but they want to keep some
attributes. This resolves the issue outlines as the problematic entry is
never emitted from the parser now.

* Fix a few issues with the XML parser

This addresses the attribute issue in the same way as the HDF5 parser,
using '.attrs' appended to the path before the attribute name. It also
adds some extra logic to only add entries if an XML tag contains
non-whitespace data.

* Update with the new convention for attributes

This updates the documentation to cover the '.attrs' convention
replacing the trailing '*' notation for attributes in formats that
support them. This change was made in the HDF5 and XML parsers.