Redesign custom parser list in Preferences dialog.
The custom parser list has been converted into a table and now displays
each custom parser directory path with the names of all the parsers
found within each directory.  Custom parser directories can now be enabled
and disabled using the checkbox.  Additionally, each parser directory's
config file can be edited by double clicking on the gear icon in a given
row of the table.

The default parser directory is now included as the first item in the
table and cannot be removed from the table.

Signed-off-by: Joey Kleingers <joey.kleingers@bluequartz.net>